requirements.txt
beautifulsoup4==4.12.0
pandas==1.5.3
plotly==5.14.1
psycopg2-binary==2.9.6
requests==2.28.2
streamlit==1.20.0
streamlit-authenticator==0.2.1
urllib3==1.26.15